Rename GCEEndpointsController to IPv4PerPod to more accurately describe the behavior.
