+++ pkg/service/ipv4perpod/ipv4perpod_test.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package gce
+package ipv4perpod
 
 import (
 	"fmt"
